Remove useless source
